package/linux-firmware: Add SafeXcel Crypto Engine 'minifw'
Install the 'minifw' firmware files for the EIP97 and EIP197 crypto
engines found in Marvell Armada 37xx, 7k, 8k, and CN913x SoCs.

Note, the firmware may be 'mini' but is written to suit the Linux
driver, as mentioned by the developer here[1].  The license is in
the WHENCE file.

[1]: https://lore.kernel.org/linux-crypto/CY4PR0401MB3652D787A7DE6E19E937A392C3B59@CY4PR0401MB3652.namprd04.prod.outlook.com/

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>